build: no entry function when suspended
The canonical way to know if a snapshot is suspended is the existence of
gate.stack section, but since it's not known yet during BindFunctions,
check the existence of gate.export.  Its purpose is to hide entry
functions for suspended snapshots, so it's the same thing in practice:
if the module contains a standard export section in addition to the
wrapper, the build would have failed with duplicate section error as
gate.export is unwrapped and parsed as a standard section.